pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imgannotator"
version = "0.1.0"
description = "CLI for applying image filters and embedding EXIF metadata"
readme = "README.md"
keywords = ["cli", "typer", "image", "exif"]
authors = [{name="Terence Schumacher", email="terenceschumacher@gmail.com"}]
license = {text="MIT"}

[project.urls]
"Homepage" = "https://github.com/tazzledazzle/imgannotator"

[project.scripts]
imgannotator = "imgannotator.cli:app"


[project.optional-dependencies]
dev = [
  "black",
  "flake8",
  "flake8-black",
  "pytest",
  "mkdocs",
  "piexif",
  "typer",
  "pillow",
  "cv2"
]
docs = [
  "mkdocs",
  "mkdocs-material",
  "mkdocs-git-revision-date-localized-plugin",
  "mkdocs-minify-plugin",
  "mkdocs-awesome-pages-plugin",
]